requirements.txt
django==5.0.0
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
pillow==10.0.0
